5555-Camera_not_working_on_fp3
issue: e/backlog#5555

Camera app has loading issue for fp3-q build devices. This issue seems
to resolved when loading cameraController in mainThread.

This commit read the `camera.load.sync` property from system. If it
present & true then it will load camera in the main thread. This
property will be set in the
https://gitlab.e.foundation/e/devices/android_device_fairphone_FP3/-/blob/v1-q/system.prop